+++ src/main/java/com/example/toy_trello/domain/comment/controller/CommentController.java
@@ -1,21 +1,17 @@
-package com.example.toy_trello.domain.comment;
+package com.example.toy_trello.domain.comment.controller;
 
+import com.example.toy_trello.domain.comment.service.CommentService;
 import com.example.toy_trello.domain.comment.dto.CommentRequestDto;
 import com.example.toy_trello.domain.comment.dto.CommentResponseDto;
-import com.example.toy_trello.domain.comment.dto.PageDto;
 import com.example.toy_trello.global.dto.CommonResponseDto;
 import com.example.toy_trello.global.security.UserDetailsImpl;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 //import org.springframework.security.core.annotation.AuthenticationPrincipal;
 //import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
-
-import javax.print.attribute.HashPrintJobAttributeSet;
-import java.util.List;
 
 @RequiredArgsConstructor
 @RequestMapping("/api/v1/comments")
